requirements.txt
mlflow
boto3
pandas
scikit-learn
numpy
matplotlib
seaborn
scipy
apache-airflow
psycopg2-binary
acryl-datahub[airflow]
sagemaker
fastapi
streamlit
awscli
joblib
mlflow[sagemaker]